requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
transformers>=4.30.0
sentence-transformers
faiss-cpu>=1.7.4
pandas>=2.0.0
numpy>=1.24.0
networkx>=3.1
scikit-learn>=1.2.0
rouge-score
tqdm>=4.65.0
ipywidgets
ipykernel
jupyter
python-dotenv
pyarrow
scispacy
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_md-0.5.1.tar.gz
neo4j>=5.8.0
matplotlib>=3.7.0